lsm: split the notifier code out into lsm_notifier.c
jira KERNEL-1156
cve CVE-2026-46054
Rebuild_History Non-Buildable kernel-6.12.0-211.22.1.el10_2
commit-author Paul Moore <paul@paul-moore.com>
commit a5e7c17c810052e94dae36f1a976a052f4490458

In an effort to decompose security/security.c somewhat to make it less
twisted and unwieldy, pull out the LSM notifier code into a new file
as it is fairly well self-contained.

No code changes.

(cherry picked from commit a5e7c17c810052e94dae36f1a976a052f4490458)
